Refactor phone collection flow into common module
