+++ src/app.ts
@@ -38,8 +38,7 @@ import { keyboardGuard } from './bot/middlewares/keyboardGuard';
 import { stateGate } from './bot/middlewares/stateGate';
 import { unknownHandler } from './bot/middlewares/unknown';
 import { callbackDecoder } from './bot/middlewares/callbackDecoder';
-import { savePhone } from './bot/middlewares/askPhone';
-import { ensurePhone } from './bot/middlewares/ensurePhone';
+import { ensurePhone, savePhone } from './bot/flows/common/phoneCollect';
 import { metricsCollector } from './bot/middlewares/metrics';
 import type { BotContext } from './bot/types';
 import { config, logger } from './config';
